refactor: split request/response packages by domain, extract named types
Split monolithic files into domain-aligned files:
- requests/: auth.go, station.go, user.go, misc.go
- responses/: errors.go, common.go, auth.go, station.go, user.go

Extract named types from anonymous structs:
- PlaylistItem, AudioURL (from StationGetPlaylist)
- GenreCategory, GenreStation (from StationGetGenreStations)
